IR funziona, PWM funziona, test led sistemati

+++ battery_handler.h
@@ -5,6 +5,6 @@ void battery_init(void);
 
 // Legge la tensione della batteria in Volt dall'ADC in scan mode.
 // Applica il rapporto del divisore di tensione (3 resistenze uguali → Vbat = 3 * Vadc).
-float Battery_ReadVoltage(void);
+double Battery_ReadVoltage(void);
 
 #endif // BATTERY_HANDLER_H
